+++ cmd/service_windows.go
@@ -31,15 +31,16 @@ import (
 	"io/fs"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strings"
 
 	"github.com/Seagate/cloudfuse/common"
+	"github.com/Seagate/cloudfuse/common/log"
 	"github.com/Seagate/cloudfuse/internal/winservice"
+	"github.com/go-ole/go-ole"
+	"github.com/go-ole/go-ole/oleutil"
 
 	"github.com/spf13/cobra"
-	"golang.org/x/sys/windows"
-	"golang.org/x/sys/windows/svc"
-	"golang.org/x/sys/windows/svc/mgr"
 )
 
 type serviceOptions struct {
@@ -48,14 +49,15 @@ type serviceOptions struct {
 }
 
 const SvcName = "cloudfuse"
+const StartupName = "CloudfuseStartup.lnk"
 
 var servOpts serviceOptions
 
 // Section defining all the command that we have in secure feature
 var serviceCmd = &cobra.Command{
 	Use:               "service",
-	Short:             "Manage cloudfuse as a Windows service. This requires Administrator rights to run.",
-	Long:              "Manage cloudfuse as a Windows service. This requires Administrator rights to run.",
+	Short:             "Manage cloudfuse mounts on Windows",
+	Long:              "Manage cloudfuse mounts on Windows",
 	SuggestFor:        []string{"ser", "serv"},
 	Example:           "cloudfuse service install",
 	FlagErrorHandling: cobra.ExitOnError,
@@ -66,78 +68,50 @@ var serviceCmd = &cobra.Command{
 
 var installCmd = &cobra.Command{
 	Use:               "install",
-	Short:             "Install as a Windows service",
-	Long:              "Install as a Windows service",
+	Short:             "Installs the startup process for Cloudfuse. Requires running as admin.",
+	Long:              "Installs the startup process for Cloudfuse which remounts any active previously active mounts on startup. . Requires running as admin.",
 	SuggestFor:        []string{"ins", "inst"},
 	Example:           "cloudfuse service install",
 	FlagErrorHandling: cobra.ExitOnError,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		err := installService()
+		dir, err := os.Getwd()
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
-			return fmt.Errorf("failed to install as a Windows service [%s]", err.Error())
+			return fmt.Errorf("unable to determine location of cloudfuse binary [%s]", err.Error())
+		}
+		programPath := filepath.Join(dir, "windows-startup.exe")
+		startupPath := filepath.Join(os.Getenv("APPDATA"), "Microsoft", "Windows", "Start Menu", "Programs", "Startup", StartupName)
+		err = makeLink(programPath, startupPath)
+		if err != nil {
+			return fmt.Errorf("unable to create startup link [%s]", err.Error())
 		}
 
+		// Create the registry for WinFsp
+		err = winservice.CreateWinFspRegistry()
+		if err != nil {
+			return fmt.Errorf("error adding Windows registry for WinFSP support [%s]", err.Error())
+		}
 		return nil
 	},
 }
 
 var uninstallCmd = &cobra.Command{
 	Use:               "uninstall",
-	Short:             "Remove as a Windows service",
-	Long:              "Remove as a Windows service",
+	Short:             "Uninstall the startup process for Cloudfuse. Requires running as admin.",
+	Long:              "Uninstall the startup process for Cloudfuse. Requires running as admin.",
 	SuggestFor:        []string{"uninst", "uninstal"},
 	Example:           "cloudfuse service uninstall",
 	FlagErrorHandling: cobra.ExitOnError,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		err := removeService()
+		startupPath := filepath.Join(os.Getenv("APPDATA"), "Microsoft", "Windows", "Start Menu", "Programs", "Startup", StartupName)
+		err := os.Remove(startupPath)
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
-			return fmt.Errorf("failed to remove as a Windows service [%s]", err.Error())
+			return fmt.Errorf("failed to delete startup process [%s]", err.Error())
 		}
 
-		return nil
-	},
-}
-
-var startCmd = &cobra.Command{
-	Use:               "start",
-	Short:             "start the Windows service",
-	Long:              "start the Windows service",
-	SuggestFor:        []string{"sta", "star"},
-	Example:           "cloudfuse service start",
-	FlagErrorHandling: cobra.ExitOnError,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		err := startService()
+		// Remove the registry for WinFsp
+		err = winservice.RemoveWinFspRegistry()
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
-			return fmt.Errorf("failed to start as a Windows service [%s]", err.Error())
-		}
-
-		return nil
-	},
-}
-
-var stopCmd = &cobra.Command{
-	Use:               "stop",
-	Short:             "stop the Windows service",
-	Long:              "stop the Windows service",
-	SuggestFor:        []string{"sto"},
-	Example:           "cloudfuse service stop",
-	FlagErrorHandling: cobra.ExitOnError,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		err := stopService()
-		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
-			return fmt.Errorf("failed to stop the Windows service [%s]", err.Error())
+			return fmt.Errorf("error removing Windows registry from WinFSP [%s]", err.Error())
 		}
 
 		return nil
@@ -160,30 +134,15 @@ var mountServiceCmd = &cobra.Command{
 			return fmt.Errorf("failed to validate options [%s]", err.Error())
 		}
 
-		// Only allow mounts if the service is running
-		running, err := isServiceRunning()
-		if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-			return errors.New("this action requires admin rights")
-		}
-		if !running {
-			return fmt.Errorf("windows service is not running")
-		}
-
 		err = mountInstance()
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
 			return fmt.Errorf("failed to mount instance [%s]", err.Error())
 		}
 
-		// Add the mount to the registry so it persists on restart.
-		err = winservice.CreateRegistryMount(servOpts.MountPath, servOpts.ConfigFile)
+		// Add the mount to the JSON file so it persists on restart.
+		err = winservice.AddMountJSON(servOpts.MountPath, servOpts.ConfigFile)
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
-			return fmt.Errorf("failed to create registry entry [%s]", err.Error())
+			return fmt.Errorf("failed to add entry to json file [%s]", err.Error())
 		}
 
 		return nil
@@ -204,23 +163,16 @@ var unmountServiceCmd = &cobra.Command{
 		// Check with winfsp to see if this is currently mounted
 		ret, err := isMounted()
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
 			return fmt.Errorf("failed to validate options [%s]", err.Error())
 		} else if !ret {
 			return fmt.Errorf("nothing is mounted here")
 		}
 
-		// Remove the mount from the registry so it does not remount on restart.
-		err = winservice.RemoveRegistryMount(servOpts.MountPath)
-		// If error is something else other than an access error then ignore it, as if unmount fails
-		// then the user will be unable to unmount because the registry will complain that the key
-		//  does not exist.
+		// Remove the mount from json file so it does not remount on restart.
+		err = winservice.RemoveMountJSON(servOpts.MountPath)
+		// If error is not nill then ignore it
 		if err != nil {
-			if errors.Is(err, windows.ERROR_ACCESS_DENIED) {
-				return errors.New("this action requires admin rights")
-			}
+			log.Err("failed to remove entry from json file [%s]", err.Error())
 		}
 
 		err = unmountInstance()
@@ -234,110 +186,37 @@ var unmountServiceCmd = &cobra.Command{
 
 //--------------- command section ends
 
-// installService adds cloudfuse as a windows service.
-func installService() error {
-	exepath, err := os.Executable()
+func makeLink(src string, dst string) error {
+	runtime.LockOSThread()
+	defer runtime.UnlockOSThread()
+
+	err := ole.CoInitializeEx(0, ole.COINIT_APARTMENTTHREADED|ole.COINIT_SPEED_OVER_MEMORY)
 	if err != nil {
 		return err
 	}
-
-	scm, err := mgr.Connect()
+	oleShellObject, err := oleutil.CreateObject("WScript.Shell")
 	if err != nil {
 		return err
 	}
-	defer scm.Disconnect() //nolint
-
-	// Don't install the service if it already exists
-	service, err := scm.OpenService(SvcName)
-	if err == nil {
-		service.Close()
-		return fmt.Errorf("%s service already exists", SvcName)
-	}
-
-	service, err = scm.CreateService(SvcName, exepath, mgr.Config{DisplayName: "Cloudfuse", StartType: mgr.StartAutomatic})
+	defer oleShellObject.Release()
+	wshell, err := oleShellObject.QueryInterface(ole.IID_IDispatch)
 	if err != nil {
 		return err
 	}
-	defer service.Close()
-
-	// Create the registry for WinFsp
-	err = winservice.CreateWinFspRegistry()
+	defer wshell.Release()
+	cs, err := oleutil.CallMethod(wshell, "CreateShortcut", dst)
 	if err != nil {
 		return err
 	}
-
-	return nil
-}
-
-// removeService uninstall the cloudfuse windows service.
-func removeService() error {
-	scm, err := mgr.Connect()
+	idispatch := cs.ToIDispatch()
+	_, err = oleutil.PutProperty(idispatch, "TargetPath", src)
 	if err != nil {
 		return err
 	}
-	defer scm.Disconnect() //nolint
-
-	service, err := scm.OpenService(SvcName)
-	if err != nil {
-		return fmt.Errorf("%s service is not installed", SvcName)
-	}
-	defer service.Close()
-
-	// Remove the registry for WinFsp
-	// Ignore error if unable to find
-	_ = winservice.RemoveWinFspRegistry()
-
-	// Remove all registry entries for cloudfuse
-	// Ignore error if registry path does not exist
-	_ = winservice.RemoveAllRegistryMount()
-
-	err = service.Delete()
+	_, err = oleutil.CallMethod(idispatch, "Save")
 	if err != nil {
 		return err
 	}
-	return nil
-}
-
-// startService starts the windows service.
-func startService() error {
-	scm, err := mgr.Connect()
-	if err != nil {
-		return err
-	}
-	defer scm.Disconnect() //nolint
-
-	service, err := scm.OpenService(SvcName)
-	if err != nil {
-		return fmt.Errorf("%s service is not installed", SvcName)
-	}
-	defer service.Close()
-
-	err = service.Start()
-	if err != nil {
-		return fmt.Errorf("%s service could not be started: %v", SvcName, err)
-	}
-	return nil
-}
-
-// startService stops the windows service.
-func stopService() error {
-	scm, err := mgr.Connect()
-	if err != nil {
-		return err
-	}
-	defer scm.Disconnect() //nolint
-
-	service, err := scm.OpenService(SvcName)
-	if err != nil {
-		return fmt.Errorf("%s service is not installed", SvcName)
-	}
-	defer service.Close()
-
-	_, err = service.Control(svc.Stop)
-	if err != nil {
-		return fmt.Errorf("%s service could not be stopped: %v", SvcName, err)
-	}
-
 	return nil
 }
 
@@ -354,31 +233,6 @@ func unmountInstance() error {
 // isMounted returns if the current mountPath is mounted using cloudfuse.
 func isMounted() (bool, error) {
 	return winservice.IsMounted(servOpts.MountPath)
-}
-
-// isServiceRunning returns whether the cloudfuse service is currently running.
-func isServiceRunning() (bool, error) {
-	scm, err := mgr.Connect()
-	if err != nil {
-		return false, err
-	}
-	defer scm.Disconnect() //nolint
-
-	service, err := scm.OpenService(SvcName)
-	if err != nil {
-		return false, err
-	}
-	defer service.Close()
-
-	status, err := service.Query()
-	if err != nil {
-		return false, err
-	}
-
-	if status.State == windows.SERVICE_RUNNING {
-		return true, nil
-	}
-	return false, nil
 }
 
 // validateMountPath checks whether the mountpath is correct and does not exist.
@@ -419,8 +273,6 @@ func init() {
 	rootCmd.AddCommand(serviceCmd)
 	serviceCmd.AddCommand(installCmd)
 	serviceCmd.AddCommand(uninstallCmd)
-	serviceCmd.AddCommand(startCmd)
-	serviceCmd.AddCommand(stopCmd)
 	serviceCmd.AddCommand(mountServiceCmd)
 	serviceCmd.AddCommand(unmountServiceCmd)